Product now adds to DB with CustomerId FK.  Changed format by swapping StringBuilder with Manager for both Product and Customer.  Changed Menu to accomodate this swap.
