+++ Bangazon/Managers/AddCustomerManager.cs
@@ -12,35 +12,20 @@ using Bangazon.StringBuilders;
 
 namespace Bangazon.Managers
 {
-    public class AddCustomerManager
+    public class CustomerManager
     {
-        public static void AddCustomer(AddCustomerStringBuilder SB)
+    private DatabaseInterface _db;
+    public CustomerManager(DatabaseInterface db)
+    {
+        _db = db;
+    }
+        public int AddCustomer(Customer _custy)
         {
-        Customer New = new Customer();  
-        Console.WriteLine ("Enter customer first name");
-        Console.Write ("> ");
-        New.FirstName = Console.ReadLine();
-        Console.WriteLine ("Enter customer last name");
-        Console.Write ("> ");
-        New.LastName = Console.ReadLine();
-        Console.WriteLine ("Enter customer Street");
-        Console.Write ("> ");    
-        New.Street = Console.ReadLine();
-        Console.WriteLine ("Enter customer City");
-        Console.Write ("> ");
-        New.City = Console.ReadLine();
-        Console.WriteLine ("Enter customer State");
-        Console.Write ("> ");
-        New.State = Console.ReadLine();
-        Console.WriteLine ("Enter customer postal code");
-        Console.Write ("> ");
-        New.ZipCode = Int32.Parse(Console.ReadLine());
-        Console.WriteLine ("Enter customer phone number");
-        Console.Write ("> ");
-        New.PhoneNumber = Console.ReadLine();     
-        SB.AddCustomer(New);
-        Menus.MainMenu();
-        }      
+            string customerString = $"INSERT into Customer VALUES (null, '{_custy.FirstName}', '{_custy.LastName}', '{_custy.Street}', '{_custy.City}', '{_custy.State}', '{_custy.ZipCode}', '{_custy.PhoneNumber}');";
+            var Y = _db.Insert(customerString);
+            return Y;
+        }
+
     }
 
 }
