+++ Bangazon/Models/Product.cs
@@ -13,6 +13,7 @@ namespace Bangazon.Models
 {
     public class Product
     {
+        public int id { get; set; }
         public int CustomerId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
